Refactor validate pruning functions
